Feature: Assured Workloads Module (#10)
* fix: integrating the new assured-workload module into the 1-resman stages

Integrated the new assured-workload module into the 1-resman stages for deployment at the tenant folder level. The goal is to take a compliance regime from the tenants variable in tfvars and build out an assured workload folder and passing the folder on to the rest of the 1-resman stage transparently

* fix: CMEK encryption for the GCS buckets created in the 1-resman stage now works.

CMEK Encryption is required for deploying with the IL5 (highest) assured workload controls. This module now creates keyring and key named 'gcs' in the `iac-core-0` project as, grants the GCS service account EncryptDecrypt permissions, and configures the GCS bucket.

* fix: Integrate folder level logging configuration

Copy the logging.tf file from the the folders/ module and adapt it so that it creates the appropriate logging configuration on the folder created under the new assured workload.

* fix: version gate the versions of "terraform", "google", and "google-beta"

This Assured Workloads module has been tested with these versions, and we want to match what the original 'folders' did.

style: fixed output.tf in the assured-workloads module

* fix: Typo in the tftest

* fix: move the iac-core-0 project for the tenant under the tenants core folder

* fix: Name the assured workload to include the compliance regime.

* Integrate AW as an addition to the existing folder module

* Finally works

* Remove redundant module

* Fixups

* Back out changes to the tests/ portion since they don't work

* readme: Include instructions for Assured Workloads in README

* Fixups from PR

* Set default locations to US, we can change this back to EU if we need to before we merge it back into upstream
* `terraform fmt -recursive` on the whole thing

* fixup: Add validation for assured workloads compliance regime

* fixup: move the -iac-core-0 project under the core folder

---------

Co-authored-by: Will Kline <admin@dino-runner.darkwolfsolutions.com>