Add tenant route back into the cloud-router
We are now populating the tenant projects with subnets all pulling from the 10.200.0.0/16 subnet. We need a static route on the NGFW to point traffic from here back to the interface 2 (landing zone).
Add a firewall rule to allow ssh to the bastion host, in case of not using IAP
Comment out references to dns.googleapis.com until it's available in IL5